requirements.txt
discord.py
fuzzywuzzy
markdownify
numpy
python-dotenv
requests